+++ NEWS.md
@@ -2,6 +2,7 @@ Unreleased
 
 20260127.0 (January 27, 2026)
 
+* Updated: `shoulda-matchers` to `v7.0`.
 * Updated: `capybara_accessible_selectors` to `v0.16.0`
 
 20260105.0 (January 5, 2026)